[CHORE] #1 - 지도 파일 분리
