+++ dnd-11th-4-iOS/dnd-11th-4-iOS/Presetation/Map/Protocol/DrawMapProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  DrawMapProtocol.swift
+//  dnd-11th-4-iOS
+//
+//  Created by 황찬미 on 7/23/24.
+//
+
+import Foundation
+
+protocol DrawMapProtocol: AnyObject {
+    func drawMap()
+}
